Make pitch update smoother

+++ Scripts/Gameplay/CharacterEntity/BaseCharacterEntity_Events.cs
@@ -24,7 +24,7 @@ namespace MultiplayerARPG
         public event System.Action<int> onCurrentFoodChange;
         public event System.Action<int> onCurrentWaterChange;
         public event System.Action<byte> onEquipWeaponSetChange;
-        public event System.Action<byte> onPitchChange;
+        public event System.Action<ushort> onPitchChange;
         public event System.Action<uint> onTargetEntityIdChange;
         // Sync lists
         public event System.Action<LiteNetLibSyncList.Operation, int> onSelectableWeaponSetsOperation;
